# frontend/openapi.yaml
openapi: 3.1.0
info:
  title: Investize API
  version: 0.1.0
  description: |
    Checked-in UI contract for the endpoints consumed by the SPA.
    The backend springdoc endpoint is authoritative and this document must be
    reconciled with GET /v3/api-docs when backend DTOs change.
servers:
  - url: /api
security:
  - bearerAuth: []
paths:
...
  /v1/portfolio/{portfolioId}:
    parameters:
      - $ref: '#/components/parameters/PortfolioId'
    get:
      operationId: getPortfolio
      responses:
        '200':
          description: Portfolio
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Portfolio'
        '404':
          $ref: '#/components/responses/Error'
    put:
      operationId: updatePortfolio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PortfolioRequest'
      responses:
        '200':
          description: Updated portfolio
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Portfolio'
    delete:
      operationId: deletePortfolio
      responses:
        '204':
          description: Deleted
  /v1/portfolio/{portfolioId}/summary:
    parameters:
      - $ref: '#/components/parameters/PortfolioId'
    get:
      operationId: getPortfolioSummary
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Portfolio valuation summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioSummary'
  /v1/portfolio/{portfolioId}/classes:
    parameters:
      - $ref: '#/components/parameters/PortfolioId'
    get:
      operationId: getPortfolioClasses
      responses:
        '200':
          description: Server-computed portfolio allocation by asset class
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PortfolioClass' }
        '503':
          $ref: '#/components/responses/Error'
  /v1/portfolio/{portfolioId}/overview:
    parameters:
      - $ref: '#/components/parameters/PortfolioId'
    get:
      operationId: getPortfolioOverview
      responses:
        '200':
          description: Portfolio totals and positions aggregated by the API
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PortfolioOverview' }
        '503':
          $ref: '#/components/responses/Error'
  /v1/portfolio/{portfolioId}/classes/{classId}:
    parameters:
      - $ref: '#/components/parameters/PortfolioId'
      - name: classId
        in: path
        required: true
        schema: { type: string, enum: [equities, realEstate, etfs, bdrs, fixedIncome, treasury] }
    get:
      operationId: getPortfolioClassDetail
      responses:
        '200':
          description: Server-computed summary and positions for an asset class
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PortfolioClassDetail' }
        '404':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
  /v1/portfolio/{portfolioId}/assets/{ticker}:
    parameters:
      - $ref: '#/components/parameters/PortfolioId'
      - name: ticker
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: getPortfolioAssetDetail
      responses:
        '200':
          description: Consolidated asset position with API-computed allocations and profitability
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PortfolioAssetDetail' }
        '404':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
  /v1/portfolio/{portfolioId}/suggestions:
    parameters:
      - $ref: '#/components/parameters/PortfolioId'
    post:
      operationId: suggestContribution
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SuggestionRequest'
      responses:
        '200':
          description: Contribution simulation; does not create movements or orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Suggestion'
        '400':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
  /v1/portfolio/{portfolioId}/recommendations/calculate:
    parameters:
      - $ref: '#/components/parameters/PortfolioId'
    post:
      operationId: calculateRecommendations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecommendationCalculationRequest'
      responses:
        '200':
          description: Portfolio-scoped contribution simulation; does not create movements or orders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecommendationCalculation'
        '400':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
...
  /v1/portfolio/{portfolioId}/allocations:
    parameters:
      - $ref: '#/components/parameters/PortfolioId'
    get:
      operationId: getAllocations
      responses:
        '200':
          description: Target allocations by category and ticker
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryAllocation'
    post:
      operationId: setAllocations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CategoryAllocation'
      responses:
        '201':
          description: Allocations saved
        '400':
          $ref: '#/components/responses/Error'
    put:
      operationId: replaceAllocations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CategoryAllocation'
      responses:
        '200':
          description: Allocations replaced
        '400':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PortfolioId:
      name: portfolioId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
    Error:
      description: API error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Decimal:
      description: Exact decimal value; API may emit a JSON number until its wire contract is migrated.
      oneOf:
        - type: string
        - type: number
...
    PortfolioRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
    Portfolio:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        totalAssets: { type: integer }
    PortfolioSummary:
      type: object
      properties:
        portfolioId: { type: integer, format: int64 }
        totalInvested: { $ref: '#/components/schemas/Decimal' }
        currentValue: { $ref: '#/components/schemas/Decimal' }
        profitOrLoss: { $ref: '#/components/schemas/Decimal' }
        percentageChange: { $ref: '#/components/schemas/Decimal' }
        assetSummaries:
          type: array
          items: { $ref: '#/components/schemas/AssetSummary' }
    AssetSummary:
      type: object
      properties:
        tickerSymbol: { type: string }
        quantity: { $ref: '#/components/schemas/Decimal' }
        averagePrice: { $ref: '#/components/schemas/Decimal' }
        totalInvested: { $ref: '#/components/schemas/Decimal' }
        currentValue: { $ref: '#/components/schemas/Decimal' }
        profitOrLoss: { $ref: '#/components/schemas/Decimal' }
        percentageChange: { $ref: '#/components/schemas/Decimal' }
    PortfolioClass:
      type: object
      required: [id, name, currentPercentage, targetPercentage, currentValue, profitability, assetsCount]
      properties:
        id: { type: string }
        name: { type: string }
        currentPercentage: { $ref: '#/components/schemas/Decimal' }
        targetPercentage: { $ref: '#/components/schemas/Decimal' }
        currentValue: { $ref: '#/components/schemas/Decimal' }
        profitability: { $ref: '#/components/schemas/Decimal' }
        assetsCount: { type: integer }
        investedAmount: { $ref: '#/components/schemas/Decimal' }
        profitAmount: { $ref: '#/components/schemas/Decimal' }
    PortfolioClassDetail:
      type: object
      properties:
        summary: { $ref: '#/components/schemas/PortfolioClass' }
        assets:
          type: array
          items: { $ref: '#/components/schemas/AssetSummary' }
    PortfolioOverview:
      type: object
      required: [portfolioId, patrimony, investedAmount, profitAmount, profitPercentage, assetsCount, assets]
      properties:
        portfolioId: { type: integer, format: int64 }
        patrimony: { $ref: '#/components/schemas/Decimal' }
        investedAmount: { $ref: '#/components/schemas/Decimal' }
        profitAmount: { $ref: '#/components/schemas/Decimal' }
        profitPercentage: { $ref: '#/components/schemas/Decimal' }
        assetsCount: { type: integer }
        assets:
          type: array
          items: { $ref: '#/components/schemas/AssetSummary' }
    PortfolioAssetDetail:
      type: object
      properties:
        ticker: { type: string }
        name: { type: string }
        sector: { type: string }
        quantity: { $ref: '#/components/schemas/Decimal' }
        averagePrice: { $ref: '#/components/schemas/Decimal' }
        currentPrice: { $ref: '#/components/schemas/Decimal' }
        investedAmount: { $ref: '#/components/schemas/Decimal' }
        currentAmount: { $ref: '#/components/schemas/Decimal' }
        profitAmount: { $ref: '#/components/schemas/Decimal' }
        profitPercentage: { $ref: '#/components/schemas/Decimal' }
        allocationInClass: { $ref: '#/components/schemas/Decimal' }
        targetAllocationInClass: { $ref: '#/components/schemas/Decimal' }
        allocationInPortfolio: { $ref: '#/components/schemas/Decimal' }
        targetAllocationInPortfolio: { $ref: '#/components/schemas/Decimal' }
        dailyVariation: { $ref: '#/components/schemas/Decimal' }
    SuggestionRequest:
      type: object
      required: [amount, currency]
      properties:
        amount: { $ref: '#/components/schemas/Decimal' }
        currency: { type: string, pattern: '^[A-Z]{3}$', default: BRL }
    RecommendationCalculationRequest:
      type: object
      required: [investmentAmount]
      properties:
        investmentAmount: { allOf: [{ $ref: '#/components/schemas/Decimal' }], exclusiveMinimum: 0 }
    RecommendationCalculation:
      type: object
      properties:
        investmentAmount: { $ref: '#/components/schemas/Decimal' }
        allocatedAmount: { $ref: '#/components/schemas/Decimal' }
        remainingAmount: { $ref: '#/components/schemas/Decimal' }
        recommendations:
          type: array
          items: { $ref: '#/components/schemas/Recommendation' }
        excludedAssets:
          type: array
          items: { $ref: '#/components/schemas/ExcludedAsset' }
        disclaimer: { type: string }
    Recommendation:
      type: object
      properties:
        ticker: { type: string }
        assetName: { type: string }
        score: { $ref: '#/components/schemas/Decimal' }
        recommendationLevel: { type: string, enum: [STRONG_BUY, MODERATE_BUY, NEUTRAL, LOW_PRIORITY] }
        suggestedAmount: { $ref: '#/components/schemas/Decimal' }
        suggestedQuantity: { $ref: '#/components/schemas/Decimal' }
        currentAllocation: { $ref: '#/components/schemas/Decimal' }
        targetAllocation: { $ref: '#/components/schemas/Decimal' }
        currentPrice: { $ref: '#/components/schemas/Decimal' }
        ceilingPrice: { $ref: '#/components/schemas/Decimal' }
        dailyVariation: { $ref: '#/components/schemas/Decimal' }
        explanation:
          type: array
          items: { type: string }
        quoteProvider: { type: string }
        quoteObservedAt: { type: string, format: date-time }
    Suggestion:
      type: object
      properties:
        requestedAmount: { $ref: '#/components/schemas/Decimal' }
        allocatedAmount: { $ref: '#/components/schemas/Decimal' }
        remainingAmount: { $ref: '#/components/schemas/Decimal' }
        currency: { type: string }
        disclaimer: { type: string }
        items:
          type: array
          items: { $ref: '#/components/schemas/SuggestionItem' }
        excludedAssets:
          type: array
          items: { $ref: '#/components/schemas/ExcludedAsset' }
    SuggestionItem:
      type: object
      properties:
        tickerSymbol: { type: string }
        assetName: { type: string }
        quantity: { $ref: '#/components/schemas/Decimal' }
        unitPrice: { $ref: '#/components/schemas/Decimal' }
        estimatedValue: { $ref: '#/components/schemas/Decimal' }
        score: { $ref: '#/components/schemas/Decimal' }
        ceilingPrice: { $ref: '#/components/schemas/Decimal' }
        ceilingDistancePercentage: { $ref: '#/components/schemas/Decimal' }
        dailyChangePercent: { $ref: '#/components/schemas/Decimal' }
        currentPercentage: { $ref: '#/components/schemas/Decimal' }
        targetPercentage: { $ref: '#/components/schemas/Decimal' }
        reasons:
          type: array
          items: { type: string }
        quoteProvider: { type: string }
        quoteObservedAt: { type: string, format: date-time }
    ExcludedAsset:
      type: object
      properties:
        tickerSymbol: { type: string }
        reason: { type: string }
...
    CategoryAllocation:
      type: object
      properties:
        category: { type: string, enum: [EQUITIES, REAL_ESTATE_FUNDS, ETFS, BDRS, FIXED_INCOME, TREASURY] }
        categoryTargetPercentage: { $ref: '#/components/schemas/Decimal' }
        assetAllocations:
          type: array
          items:
            type: object
            properties:
              tickerSymbol: { type: string }
              targetPercentage: { $ref: '#/components/schemas/Decimal' }
...
    Error:
      type: object
      required: [timestamp, status, code, message, details]
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer }
        code: { type: string }
        message: { type: string }
        details:
          type: object
          additionalProperties: true